+++ mindbox_android/test/mindbox_android_test.dart
@@ -56,7 +56,7 @@ void main() {
       expect(
           () async =>
               MindboxPlatform.instance.init(configuration: invalidConfig),
-          throwsA(isA<MindboxException>()));
+          throwsA(isA<MindboxInitializeError>()));
     },
   );
 
@@ -174,6 +174,234 @@ void main() {
       ).whenComplete(() => completer.complete('invoked'));
 
       expect(completer.isCompleted, isFalse);
+    },
+  );
+
+  test(
+    'When SDK was initialized, executeSyncOperation() should be invoked',
+    () async {
+      final completer = Completer<String>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      MindboxPlatform.instance
+          .executeSyncOperation(
+            operationSystemName: 'dummy-name',
+            operationBody: {'dummy-key': 'dummy-value'},
+            onSuccess: (success) {},
+            onError: (error) {},
+          )
+          .whenComplete(() => completer.complete('invoked'));
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      expect(completer.isCompleted, isTrue);
+    },
+  );
+
+  test(
+    'When SDK not initialized, executeSyncOperation() should not be invoked',
+    () async {
+      final completer = Completer<String>();
+
+      MindboxPlatform.instance
+          .executeSyncOperation(
+            operationSystemName: 'dummy-name',
+            operationBody: {'dummy-key': 'dummy-value'},
+            onSuccess: (success) {},
+            onError: (error) {},
+          )
+          .whenComplete(() => completer.complete('invoked'));
+
+      expect(completer.isCompleted, isFalse);
+    },
+  );
+
+  test(
+    'When no errors occur during execution, executeSyncOperation() should '
+    'return success response',
+    () async {
+      final completer = Completer<String>();
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-valid-name',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (response) => completer.complete(response),
+        onError: (error) => completer.complete(error.toString()),
+      );
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      expect(await completer.future, equals('dummy-response'));
+    },
+  );
+
+  test(
+    'When validation data is incorrect, executeSyncOperation() should throw'
+    'MindboxValidationError to onError callback',
+    () async {
+      final completer = Completer<Exception>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-validation-error',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (success) {},
+        onError: (error) => completer.completeError(error),
+      );
+
+      expect(() => completer.future, throwsA(isA<MindboxValidationError>()));
+    },
+  );
+
+  test(
+    'When operation data is incorrect, executeSyncOperation() should throw'
+    'MindboxProtocolError to onError callback',
+    () async {
+      final completer = Completer<Exception>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-invalid-system-name',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (success) {},
+        onError: (error) => completer.completeError(error),
+      );
+
+      expect(() => completer.future, throwsA(isA<MindboxProtocolError>()));
+    },
+  );
+
+  test(
+    'When server returns internal error, executeSyncOperation() should throw'
+    'MindboxServerError to onError callback',
+    () async {
+      final completer = Completer<Exception>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-server-error',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (success) {},
+        onError: (error) => completer.completeError(error),
+      );
+
+      expect(() => completer.future, throwsA(isA<MindboxServerError>()));
+    },
+  );
+
+  test(
+    'When network error occurred, executeSyncOperation() should '
+    'return MindboxNetworkError to onError callback',
+    () async {
+      final completer = Completer<Exception>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-network-error',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (success) {},
+        onError: (error) => completer.completeError(error),
+      );
+
+      expect(() => completer.future, throwsA(isA<MindboxNetworkError>()));
+    },
+  );
+
+  test(
+    'When Mindbox SDK internal error occurred, executeSyncOperation() should '
+    'return MindboxInternalError to onError callback',
+    () async {
+      final completer = Completer<Exception>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-internal-error',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (success) {},
+        onError: (error) => completer.completeError(error),
+      );
+
+      expect(() => completer.future, throwsA(isA<MindboxInternalError>()));
+    },
+  );
+
+  test(
+    'When response data from native Mindbox SDK is empty or null , '
+    'executeSyncOperation() should return MindboxInternalError '
+    'to onError callback',
+    () async {
+      final completer = Completer<Exception>();
+
+      final validConfig = Configuration(
+        domain: 'domain',
+        endpointIos: 'endpointIos',
+        endpointAndroid: 'endpointAndroid',
+        subscribeCustomerIfCreated: true,
+      );
+
+      await MindboxPlatform.instance.init(configuration: validConfig);
+
+      MindboxPlatform.instance.executeSyncOperation(
+        operationSystemName: 'dummy-null-error-message',
+        operationBody: {'dummy-key': 'dummy-value'},
+        onSuccess: (success) {},
+        onError: (error) => completer.completeError(error),
+      );
+
+      expect(() => completer.future, throwsA(isA<MindboxInternalError>()));
     },
   );
 }