+++ mindbox_platform_interface/lib/src/types/types.dart
@@ -1,2 +1,3 @@
 export 'configuration.dart';
 export 'mindbox_method_handler.dart';
+export 'mindbox_mock_method_call_handler.dart';