+++ mindbox/CHANGELOG.md
@@ -1,7 +1,13 @@
+## 1.0.0
+
+* Stable release of 'mindbox' plugin. Features:
+  * receive and show push notification in both mobile platforms.
+  * receive push notification data(links) in Flutter.
+  * execute sync/async operations.
+
 ## 0.2.0
 
 * Add execute async operation.
-
 
 ## 0.1.0