+++ mindbox_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0
+
+* Add execute sync operation.
+
 ## 0.2.0
 
 * Add execute async operation.